+++ src/schema/assets/asset.rs
@@ -3,14 +3,14 @@
 
 use serde::{Deserialize, Serialize};
 
+use crate::schema::helpers::visual_object::VisualObject;
+
 ///
 #[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
 pub struct Asset {
     /// Unique identifier used by layers when referencing this asset
     #[serde(rename = "id")]
     pub id: String,
-    /// Human readable name
-    #[serde(rename = "nm")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
+    #[serde(flatten)]
+    pub visual_object: VisualObject,
 }