+++ src/schema/assets/image.rs
@@ -1,6 +1,8 @@
 // Copyright 2024 the Velato Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+use crate::schema::helpers::slottable_object::SlottableObject;
+
 use super::file_asset::FileAsset;
 use serde::de::Deserializer;
 use serde::{Deserialize, Serialize, Serializer};
@@ -8,8 +10,12 @@ use serde::{Deserialize, Serialize, Serializer};
 /// External image
 #[derive(Deserialize, Serialize, Debug, Clone, PartialEq)]
 pub struct Image {
+    /// File asset properties
     #[serde(flatten)]
     pub file_asset: FileAsset,
+    /// Slottable object properties
+    #[serde(flatten)]
+    pub slottable_object: SlottableObject,
     /// Width of the image
     #[serde(rename = "w")]
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -42,7 +48,6 @@ where
 }
 
 // Function signature must match Serde's Serialize trait, so need to suppress Clippy.
-#[expect(clippy::trivially_copy_pass_by_ref, reason = "Deferred")]
 pub fn seq_to_str<S>(v: &Option<bool>, serializer: S) -> Result<S::Ok, S::Error>
 where
     S: Serializer,
@@ -59,7 +64,9 @@ mod tests {
     use super::Image;
     use crate::schema::{
         assets::{asset::Asset, file_asset::FileAsset},
-        helpers::int_boolean::BoolInt,
+        helpers::{
+            int_boolean::BoolInt, slottable_object::SlottableObject, visual_object::VisualObject,
+        },
     };
     use once_cell::sync::Lazy;
     use serde_json::json;
@@ -76,10 +83,14 @@ mod tests {
         )
     });
     static IMAGE: Lazy<Image> = Lazy::new(|| Image {
+        slottable_object: SlottableObject::default(),
         file_asset: FileAsset {
             asset: Asset {
                 id: "my image".to_string(),
-                name: None,
+                visual_object: VisualObject {
+                    name: None,
+                    match_name: None,
+                },
             },
             dir: None,
             file_name: "data:image/png;base64,...".to_string(),